requirements.txt
Django==4.2.13
PyJWT==2.8.0
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
defusedxml==0.8.0rc2
djangorestframework==3.15.1
idna==3.7
importlib-metadata==7.1.0
kombu==5.3.7
oauthlib==3.2.2
pillow==10.3.0
pip==24.0
prompt-toolkit==3.0.43
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
python-social-auth==0.3.6
python3-openid==3.2.0
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
segno==1.6.1
setuptools==69.1.1
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0	5.1.0
wcwidth==0.2.13
wheel==0.42.0
zipp==3.18.1